third_party/blink/web_tests/platform/generic/fast/box-sizing/css-table-with-box-sizing-expected.txt

Test case for bug 89819. This tests CSS 'table' and 'inline-table' with box-sizing.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementById("t1").offsetWidth is 120
PASS document.getElementById("t1").offsetHeight is 120
PASS document.getElementById("t2").offsetWidth is 120
PASS document.getElementById("t2").offsetHeight is 120
PASS document.getElementById("t3").offsetWidth is 140
PASS document.getElementById("t3").offsetHeight is 140
PASS document.getElementById("t4").offsetWidth is 140
PASS document.getElementById("t4").offsetHeight is 140
PASS successfullyParsed is true

TEST COMPLETE
